Move pangoText example from example script to Docs (#533)
* moved pangoText example to docs

* Update text_mobject.py

* Update text_mobject.py

fix docsting
